Visualizing Principal Component Analysis

The PCAviz package provides a simple interface for quickly creating visually compelling plots from Principal Component Analysis (PCA) and accompanying data. See the vignettes included in this package for examples of plots generated with the PCAviz package. The plotting functions are also suitable for other dimensionality reduction techniques such as Multidimensional Scaling (MDS) and Local Linear Embedding (LLE).

Introduction

The “iris”, “popres” and “regmap” vignettes give a quick overview of the package's key features and functions.

PCAviz is specifically designed for visualizing three types of information jointly: (1) PCA results; (2) continuous covariates such as geographical co-ordinates; and (3) categorical data such as group labels. The pcaviz S3 class combines all these data stored into a single object. See pcaviz-plots for details on the PCAviz plotting interface.

The main functions operate on pcaviz objects. Most functions in the PCAviz package start with pcaviz_, so a list of available PCAviz functions can be quickly scanned using tab completion. For a complete list of help pages, use help(package = "PCAviz").

See also

http://github.com/NovembreLab/PCAviz

Examples

## Not run: ------------------------------------
# # Demonstrations of the PCAviz package.
# vignette("iris")
# vignette("popres")
# vignette("regmap")
## ---------------------------------------------