Principal components analysis of genetic variation in POPRES sample

Data on 1,387 European subjects from the POPRES panel (Nelson et al, 2008), and results from principal components analysis (PCA) of genotypes at 197,146 single nucleotide polymorphisms (SNPs).

data(popres)

Format

popres$x is a data frame with 1,387 rows and 10 columns, in which each row corresponds to an individual:

id

sample id

country

country of origin

abbrv

abbreviated country label

color

mapping from countries to colors

longitude

geographic longitude, in degrees

latitude

geographic latitude, in degrees

PC1--4

projection of sample onto first 4 PCs

popres$rotation is a data frame containing the eigenvectors ("loadings") computed from the genotype data at a subset of the SNPs. The first two columns give the chromosome number (1--22) and base-pair position of each SNP.

Details

Note that some of the country designations do not reflect current political boundaries.

Also note that "country" and "abbrv" columns do not correspond exactly; for example, abbreviations "GB" (Great Britain) and "Sct" (Scotland) are both assigned to samples from the United Kingdom.

Geographic coordinates are the same for each country, so all samples with the same country of origin are assigned the same latitude and longitude.

Source

Data were retrieved and compiled from http://github.com/NovembreLab/Novembre_etal_2008_misc. To reduce file size, only the top 4 PCs were retained.

References

Nelson, M. R., Bryc, K., King, K. S., Indap, A., Boyko, A. R., Novembre, J., et al (2008). The population reference sample, POPRES: a resource for population, disease, and pharmacological genetics research. American Journal of Human Genetics 83, 347--358. Novembre, J., Johnson, T., Bryc, K., Kutalik, Z., Boyko, A. R., Auton, A., Indap, A., King, K. S., Bergmann, S., Nelson, M. R., Stephens, M. and Bustamante, C. D. (2008) Genes mirror geography within Europe. Nature 456, 274-–274.

Examples

# View the first few samples and genetic markers.
data(popres)
head(popres$x[c(1:3,7:8)])

#>     id                country abbrv     PC1     PC2
#> 1  474                Croatia    HR  0.0111 -0.0460
#> 2  660  Serbia and Montenegro    RS -0.0300 -0.0443
#> 3 1890         Czech Republic    CZ  0.0137 -0.0444
#> 4 1923 Bosnia and Herzegovina    BA  0.0028 -0.0493
#> 5 2083  Serbia and Montenegro    RS -0.0031 -0.0416
#> 6 2733  Serbia and Montenegro    RS -0.0026 -0.0424

head(popres$rotation)

#>               chr     pos    PC1    PC2    PC3    PC4
#> SNP_A-1909444   1  742429 -0.860  0.026  0.536  0.150
#> SNP_A-1886933   1  775852 -1.140 -0.554  0.511  0.879
#> SNP_A-2116190   1  993492  0.230 -0.585  1.338 -0.065
#> SNP_A-4291020   1 1087198 -0.469  0.396 -0.835 -0.466
#> SNP_A-1902458   1 1120590  1.606 -0.912  0.214  0.315
#> SNP_A-2131660   1 1145994 -0.327  0.905  0.832  1.637

## Not run: ------------------------------------
# # See the vignette for more examples.
# vignette("popres")
## ---------------------------------------------