Principal components analysis of A. thaliana RegMap data

Data on 1,307 Arabidopsis thaliana accessions collected around the world. Data include geographic co-ordinates and results of applying principal components analysis (PCA) to genotypes at 214,051 single nucleotide polymorphisms (SNPs).

data(regmap)

Format

regmap is a data frame with 1,307 rows and 21 columns, in which each row represents an A. thaliana accession included in the RegMap panel. The last 11 columns (labeled "PC1" through "PC10") store the projection of the samples onto the first 10 PCs.

Source

Arabidopsis thaliana SNP genotypes were obtained from the website of Joy Bergelson's lab at the University of Chicago, http://bergelson.uchicago.edu/?page_id=790.

Details

PCs were computed using rpca from the rsvd package after first converting the genotypes to a 1,307 x 214,051 binary matrix.

References

Horton, M. W., Hancock, A. M., Huang, Y. S., Toomajian, C., Atwell, S., Auton, A., et al (2012). Genome-wide patterns of genetic variation in worldwide Arabidopsis thaliana accessions from the RegMap panel. Nature Genetics 44, 212–216.

Examples

# View the first few data samples.
data(regmap)
head(regmap)

#>   array_id ecotype_id median_intensity latitude longitude nativename firstname
#> 1      641       9381              476    55.81     14.21      Fri 1    Alison
#> 2      642       9380              556    55.75     13.37     FlyA 3    Alison
#> 3      643       9378              542    63.02     18.32      FaU 4    Alison
#> 4      644       9371              478    63.02     18.32      FaL 1    Alison
#> 5      645       9367              560    57.68     15.00      EkN 3    Alison
#> 6      646       9363              363    62.91     18.40      EdJ 2    Alison
#>     surname            site   region country    PC1     PC2    PC3    PC4
#> 1 Anastasio       Frisaboda S Sweden     SWE -27.76  34.484  6.258 -2.040
#> 2 Anastasio         Flyinge S Sweden     SWE -20.18  34.954 15.606 -6.110
#> 3 Anastasio  Faberget-Upper N Sweden     SWE -56.76  89.701  9.092 11.483
#> 4 Anastasio  Faberget-Lower N Sweden     SWE -55.64  88.580  8.069  9.981
#> 5 Anastasio Eksjoholm-Norra C Sweden     SWE -38.55   6.197 -6.996 -5.794
#> 6 Anastasio    Edsattersjon N Sweden     SWE -60.68 109.877 16.130 13.553
#>       PC5     PC6      PC7    PC8     PC9    PC10
#> 1  -6.575  -1.444   2.2803  25.62 -10.911  18.614
#> 2 -99.281 -44.441 -38.4179 -40.01  24.640   6.548
#> 3  46.634  -1.500   0.0836 -14.24  -3.090  -6.131
#> 4  46.185  -1.223   1.5694 -15.35  -2.364  -4.185
#> 5 -12.844   1.417   5.9657  23.06  -3.204 -23.093
#> 6  46.365  -2.152 -10.4390 -20.77  10.236  -1.030

## Not run: ------------------------------------
# # See the vignette for more examples.
# vignette("regmap")
## ---------------------------------------------